!!!MAJOR CHANGES IN SVN FOLDER STRUCTURE!!!
because there are in fact three separate projects in our trunk (editor, backend/repository and oryx-on-a-stick) I've rearranged the files in the repository. the main folders in the trunk are
 - poem-jvm (backend)
 - editor (the editor; client and servlets)
 - stick (oryx-on-a-stick)

with this structure we can easily add additional projects to the repository.

the files that belongs to the editor itself are now all moved to the folder 'editor'.

but there is only one ant build file that controls all projects. this is because all projects rely on editor so that it makes no sense to seperate them. the build script is adjusted for the new folder structure. 

there is one important change in the build file to consider: if you want to deploy the xhtml test files, too (include them in the war file) than you have to call the target deploy-xhtml before building the war file. otherwise the xhtml files are not deployed.


git-svn-id: http://oryx-editor.googlecode.com/svn/trunk@1216 d672c736-503d-0410-a38a-9366997c882b