Fixes for BPMSPL-102 (details below)
- Don't save on reload when process is locked by a different
user. Saving would override the lock owners changes and also
unlock the process. This fix was in place to some degree
but failed to consider that the lock status could have
changed since the editor/process was opened.

- Allow to switch between edit and read-only mode. When a
process is locked the editor will open in read-only mode.
If the process is then unlocked there was previously no
way for already open editors to enter edit mode although
they were able to acquire the lock. This happens
automatically now.