JBPM-8233 fixed compilation and tests for jbpm designer
For backend:
java.xml.bind (JAXB) - REMOVED in java 11, dependency added:
javax.xml.bind:jaxb-api

For client:
Removed dependency to powermock since the only way to compile it with
java 11 is upgrade version to 2. But after that a lot of test failing.
(With high probability to break a lot of tests across the entire KIE
Group projects suite).

To achieve this small refactoring incorporated.
1. Allows to mock two methods in StringUtils class (removed static)
2. Ability to replace basic StringUtils by mock in Assignments class.